pyproject.toml
[project]
name = "llmmj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.54.0",
    "evals>=3.0.1.post1",
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.4",
    "google-adk>=1.5.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.15",
    "langchain-community>=0.3.25",
    "langchain-deepseek>=0.1.3",
    "langchain-google-genai>=2.0.10",
    "langchain-mcp>=0.2.1",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.23",
    "litellm>=1.73.6",
    "mahjong>=1.3.0",
    "mcp>=1.9.4",
    "openai>=1.86.0",
    "openai-agents[litellm]>=0.1.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "pytest>=8.4.0",
    "ruff>=0.12.1",
]
